fix(createP2PIndexesUpdatedInvariant): move common to utils
